Abstract
Health has always been a concern of people, and there are many factors that affect our health. Some of these factors are attributes that people can control, such as smoking or exercising; others are disabilities or life status are ones that are not controllable, but still affect human health. This project focuses on one specific health concern of humans, which is having a stroke. Stroke is ranked as the second leading cause of death worldwide with an annual mortality rate of about 5.5 million (Donkor, 2018). This study investigates how the possibility of having a stroke is explained by external, internal, and social attributes. Firstly, the physical conditions of the individuals such as BMI and age were tested to find a relationship. As a result, it was found that the median BMIs of the people who have had strokes are higher. Moreover, it was also found that their ages, independent from the first result, are also higher. Secondly, external factors such as work type, residence type, smoking and marital status were examined. A weak positive relationship between work type and having a stroke was found. It can be deduced that self-employed people are more likely to have a stroke. Again, a weak negative relationship was found between smoking status and stroke. However, it is negligible; therefore, it can be said that smoking status has no relationship with having a stroke. Another factor that was examined was marriage status, a strong relationship was found. It can be concluded that married people are 4.5 times more likely to have a stroke. Then the past and current diseases and health conditions such as heart disease, hypertension, and glucose level were examined to find a relationship. It was found that people with heart diseases are 5.24 times more likely to get a stroke and people with hypertension are 4.43 times more likely to get a stroke. There was a positive relationship between stroke and glucose level, it may be confirmed that people who had a stroke and who did not have a stroke do not have the same median glucose levels. To sum up, it can be concluded that present or past diseases and / or disorders, such as heart disease, hypertension, sub-optimal glucose level increase the likelihood of having a stroke. Finally, selected covariates were gathered to conduct a model to predict stroke status. The conducted model has a binomial response for stroke and has the covariates age, hypertension, heart disease status, and log of average glucose level. According to the model; when all the covariates were held constant; the odds of having a stroke increases by 39.4 % with each additional age of the patients. The odds of having a stroke increase by 63.1 % if the patient has hypertension. Similarly, the odds of having a stroke increase by 55.1 % if the patient has heart disease. Lastly, the odds of having a stroke increases by 37 % with each additional average level of glucose of the patients.